fix cache clearing race condition in unit tests
